refactor(experiments): expose summary contracts via the facade
